Fixed multiple casting warnings Removed afxres.h as it is not needed and is also a default header file provided by VS.

+++ 0.56.b/graysvr/CCharSpell.cpp
@@ -2709,7 +2709,7 @@ reflectit:
 				iDmg *= ((pCharSrc->Skill_GetBase(SKILL_EVALINT) * 3) / 1000) + 1;
 
 				// Spell Damage Increase bonus
-				int DamageBonus = pCharSrc->GetDefNum("INCREASESPELLDAM",true);
+				int DamageBonus = static_cast<int>(pCharSrc->GetDefNum("INCREASESPELLDAM",true));
 				if ( m_pPlayer && pCharSrc->m_pPlayer && DamageBonus > 15 )		// Spell Damage Increase is capped at 15% on PvP
 					DamageBonus = 15;
